Change the contract of `ExecutionPlan.selectNext()` so that it only returns nodes that need to be executed.
Previously it would return nodes that are already complete, for example if a task in an included build happened to already run to build some plugins, and nodes that must not run because their dependencies  failed.

Do not add nodes to the plan at all if they have already completed at the time the plan is assembled.